+++ src/main/java/com/kustacks/kuring/common/utils/validator/BadWordChecker.java
@@ -1,0 +1,5 @@
+package com.kustacks.kuring.common.utils.validator;
+
+public interface BadWordChecker {
+    void process(String content);
+}
