+++ src/main/java/com/kustacks/kuring/common/utils/validator/BadWordInitProcessor.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.notice.application.port.in;
+package com.kustacks.kuring.common.utils.validator;
 
 public interface BadWordInitProcessor {
     void process();
