+++ src/main/java/com/kustacks/kuring/notice/application/service/NoticeCommandService.java
@@ -1,33 +1,25 @@
 package com.kustacks.kuring.notice.application.service;
 
 import com.kustacks.kuring.common.annotation.UseCase;
-import com.kustacks.kuring.common.exception.BadWordContainsException;
 import com.kustacks.kuring.common.exception.NoPermissionException;
 import com.kustacks.kuring.common.exception.NotFoundException;
 import com.kustacks.kuring.common.exception.code.ErrorCode;
-import com.kustacks.kuring.notice.application.port.in.BadWordInitProcessor;
+import com.kustacks.kuring.common.utils.validator.BadWordChecker;
 import com.kustacks.kuring.notice.application.port.in.NoticeCommentDeletingUseCase;
 import com.kustacks.kuring.notice.application.port.in.NoticeCommentEditingUseCase;
 import com.kustacks.kuring.notice.application.port.in.NoticeCommentWritingUseCase;
-import com.kustacks.kuring.notice.application.port.out.BadWordsQueryPort;
 import com.kustacks.kuring.notice.application.port.out.CommentCommandPort;
 import com.kustacks.kuring.notice.application.port.out.CommentQueryPort;
 import com.kustacks.kuring.notice.application.port.out.NoticeQueryPort;
 import com.kustacks.kuring.notice.application.port.out.dto.CommentReadModel;
 import com.kustacks.kuring.notice.application.port.out.dto.NoticeDto;
-import com.kustacks.kuring.notice.domain.BadWord;
 import com.kustacks.kuring.notice.domain.Comment;
 import com.kustacks.kuring.user.application.port.out.RootUserQueryPort;
 import com.kustacks.kuring.user.domain.RootUser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.ahocorasick.trie.Emit;
-import org.ahocorasick.trie.Trie;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.PostConstruct;
-import java.util.Collection;
-import java.util.List;
 import java.util.Objects;
 
 @Slf4j
@@ -37,25 +29,16 @@ import java.util.Objects;
 public class NoticeCommandService implements
         NoticeCommentWritingUseCase,
         NoticeCommentEditingUseCase,
-        NoticeCommentDeletingUseCase,
-        BadWordInitProcessor {
-    private static final String COMMENT_REGEX = "[^가-힣a-zA-Z0-9]";
+        NoticeCommentDeletingUseCase {
     private final NoticeQueryPort noticeQueryPort;
     private final CommentCommandPort commentCommandPort;
     private final CommentQueryPort commentQueryPort;
     private final RootUserQueryPort rootUserQueryPort;
-    private final BadWordsQueryPort badWordQueryPort;
-
-    private Trie badWordTrie;
-
-    @PostConstruct
-    public void badWordInit() {
-        process();
-    }
+    private final BadWordChecker badWordChecker;
 
     @Override
     public void process(WriteCommentCommand command) {
-        validateText(command.content());
+        badWordChecker.process(command.content());
 
         RootUser rootUser = findRootUserByEmailOrThrow(command.email());
 
@@ -68,7 +51,7 @@ public class NoticeCommandService implements
 
     @Override
     public void process(WriteReplyCommand command) {
-        validateText(command.content());
+        badWordChecker.process(command.content());
 
         RootUser rootUser = findRootUserByEmailOrThrow(command.email());
 
@@ -88,7 +71,7 @@ public class NoticeCommandService implements
 
     @Override
     public void process(EditCommentCommand command) {
-        validateText(command.content());
+        badWordChecker.process(command.content());
 
         RootUser rootUser = findRootUserByEmailOrThrow(command.email());
 
@@ -120,35 +103,6 @@ public class NoticeCommandService implements
         commentCommandPort.delete(findComment);
 
         log.info("delete notice-comment, user{}, notice{}, comment{}", rootUser.getId(), findNotice.getId(), findComment.getId());
-    }
-
-    @Override
-    public void process() {
-        List<BadWord> activeBadWords = badWordQueryPort.findAllByActive();
-        if (!activeBadWords.isEmpty()) {
-            Trie.TrieBuilder builder = Trie.builder().ignoreCase();
-
-            for (BadWord badWord : activeBadWords) {
-                builder.addKeyword(badWord.getWord());
-            }
-
-            badWordTrie = builder.build();
-        }
-
-        log.info("금칙어 로드 완료 - 총 {} 개", activeBadWords.size());
-    }
-
-    private void validateText(String content) {
-        if (badWordTrie != null) {
-            // 특수문자 제거 및 소문자 변환
-            String normalizedContent = content.replaceAll(COMMENT_REGEX, "").toLowerCase();
-
-            // 전체 텍스트에 대해 금칙어 검사
-            Collection<Emit> matches = badWordTrie.parseText(normalizedContent);
-            if (!matches.isEmpty()) {
-                throw new BadWordContainsException(ErrorCode.COMMENT_BAD_WORD_CONTAINS);
-            }
-        }
     }
 
     private static boolean isNotCommentOwner(Comment findComment, RootUser findUser, NoticeDto findNotice) {
